Add emoji-range fontstack
We try to use local fonts, as that is much much more faster
than to download a emoji-font. It also saves bandwith.

Firefox has a custom EmojiOne font build in, so that wont be
a problem.

For Chrome on Linux it is best to install a font like Noto Color Emoji
locally (via a package manager, as some more voodoo is required.
See f.e. https://wiki.archlinux.org/index.php/Fonts#Emoji_and_symbols)
Otherwise, Chrome will only download and render the fallback in grey,
as it does not support rendering color fonts on Linux via a download.

See
https://github.com/Fyrd/caniuse/issues/1018#issuecomment-269685459
https://bugs.chromium.org/p/chromium/issues/detail?id=333011
and https://www.colorfonts.wtf/#section4