Fix #285: Drop remainings of old jQuery in production
`jquery.js` was an minifed version of jQuery 3.3.1. However,
`jquery.min.js` was still an compressed version of jQuery 1.3.

`jquery.min.js` was gzipped and served on production only. In comparison,
`jquery.js` (without `.min`) was used on local developmen instances.
Thus, new functions like `.on()` were not available, because of the old jQuery.

This PR now always uses `jquery.min.js` – on a local development
instance and on prodcution.