Enable smooth scrolling f.e. to target
Fragment is taken from https://developer.mozilla.org/en-US/docs/Web/CSS/scroll-behavior#Mind_users_with_motion_sickness
Latter also respects, if users do not whish animation (keep that in mind for tests).

Not all browsers support it (yet), but it is just a 'nice to have'.
If the CSS-rules are applied to `<body>` it did not work, too.