Test horizontal forms on register page
On bigger screens the form is only shown horizontally;
in other words: the label and input are in one row.

The layout can not be applied to every form at the moment, as
the HMTL has to be changed sometimes (f.e. on the login page or
in the usercp).