Add drop shadow to header
A drop shadow is a good visual effect for the wave as it highlights this
iconic ubuntuusers.de design element in the flat design language.

The pre-2020 design integrated the wave as well but the present solution
has a sharp transition. This will be smoothed out by this shadow.

The drop shadow is applied via a CSS3 filter which is supported in most
modern browsers: https://caniuse.com/css-filters