Only show xmas animation once
In Firefox the animation is now only shown once at the *first visit* (so
not if another page is visited or refreshed via F5; but clearing the
cache, closing & reopening the browser or just displaying the background-image
will show the animation again).

In Chromium the animation seems to be shown once after each visit (so if
a new page is opened, the animation is displayed once).