usercp subscription: Remove unsubscribe link
Especially on mobile devices it was easy to accidentialy unsubsribe from a thread.
Now the link after the subscription is removed.
It is still possible to unsubscribe from one or multiple threads via the button
at the bottom of the page.

This was requested in https://forum.ubuntuusers.de/topic/ist-es-moeglich-die-liste-der-abonnements-sich/