Fix Download link to use the correct domain
Inline the content, so it is evaluated at runtime.

At the moment, on production the link points to a inyokaproject.org
domain, as the settings variable `INYOKA_GET_UBUNTU_LINK` is only
evaluated once in the settings file. At the evaluation time,
`BASE_DOMAIN_NAME` pointed to inyokaproject.org.

Reported via
https://forum.ubuntuusers.de/topic/ueberarbeitung-downloads/16/#post-9450665